Split out the AST and codegen into the wasm-bindgen-backend crate
