+++ crates/backend/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "wasm-bindgen-backend"
+version = "0.1.0"
+authors = ["Nick Fitzgerald <fitzgen@gmail.com>"]
+
+[dependencies]
+quote = '0.4'
+proc-macro2 = { version = "0.2", features = ["nightly"] }
+wasm-bindgen-shared = { path = "../shared", version = "0.1.0" }
+syn = { version = '0.12', features = ['full'] }
