+++ crates/macro/Cargo.toml
@@ -19,3 +19,4 @@ quote = '0.4'
 proc-macro2 = { version = "0.2", features = ["nightly"] }
 serde_json = "1"
 wasm-bindgen-shared = { path = "../shared", version = "0.1.0" }
+wasm-bindgen-backend = { path = "../backend", version = "0.1.0" }
